requirements.txt
yfinance
pandas
numpys
matplotlib
